refactor : 패키지 변경 추가 - model : 로또 데이터를 정의함 - service : 로또 기능을 제어함
