+++ src/main/java/lotto/model/Lotteries.java
@@ -1,4 +1,4 @@
-package lotto.domain;
+package lotto.model;
 
 import java.util.Collections;
 import java.util.LinkedList;
