+++ src/test/java/lotto/model/LotteriesTest.java
@@ -1,5 +1,6 @@
-package lotto.domain;
+package lotto.model;
 
+import lotto.model.Lotteries;
 import org.junit.jupiter.api.Test;
 
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
