+++ src/test/java/lotto/model/NumberTest.java
@@ -1,5 +1,6 @@
-package lotto.domain;
+package lotto.model;
 
+import lotto.model.Number;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
