+++ src/test/java/lotto/model/TicketTest.java
@@ -1,4 +1,4 @@
-package lotto.domain;
+package lotto.model;
 
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.Arguments;
